Split widget: batch list
